+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   - Process It blocks in memory [GH-123]
   - Fixed -ExecutionPolicy in pester.bat [GH-130]
   - Add support for mocking internal module functions, aliases, exe and filters. [GH-126]
+  - Fix TestDrive clean up [GH-129]
 
 ## 2.0.4 (March 9, 2014)